modified:   app/controllers/application_controller.rb 	new file:   app/controllers/parents/registrations_controller.rb 	new file:   app/controllers/parents/sessions_controller.rb 	modified:   app/controllers/parents_controller.rb 	modified:   app/controllers/sessions_controller.rb 	new file:   app/controllers/users_registrations_controller.rb 	modified:   app/helpers/application_helper.rb 	new file:   app/helpers/parents/registrations_helper.rb 	new file:   app/helpers/users/registrations_helper.rb 	modified:   app/models/parent.rb 	modified:   app/models/user.rb 	modified:   app/policies/admin_dashboard_policy.rb 	modified:   app/policies/admin_policy.rb 	modified:   app/policies/application_policy.rb 	modified:   app/policies/check_in_policy.rb         modified:   app/policies/parent_policy.rb 	modified:   app/policies/user_policy.rb 	modified:   app/views/admin/dashboard.html.erb 	modified:   app/views/children/new.html.erb 	modified:   app/views/devise/sessions/new.html.erb 	modified:   app/views/devise/shared/_links.html.erb 	modified:   app/views/home/index.html.erb 	modified:   app/views/layouts/application.html.erb 	deleted:    app/views/parents/new.html.erb 	new file:   app/views/parents/registrations/new.html.erb 	modified:   app/views/parents/show.html.erb 	modified:   config/environments/development.rb 	modified:   config/routes.rb 	new file:   db/migrate/20241021161238_add_user_id_to_parents.rb 	new file:   db/migrate/20241027202335_add_confirmed_at_to_users.rb 	new file:   db/migrate/20241028015611_add_devise_fields_to_parents.rb 	new file:   db/migrate/20241028023420_add_trackable_columns_to_parents.rb 	modified:   db/schema.rb 	new file:   test/controllers/parents/registrations_controller_test.rb 	new file:   test/controllers/users/registrations_controller_test.rb

+++ app/controllers/parents_controller.rb
@@ -1,5 +1,4 @@
 class ParentsController < ApplicationController
-  include Pundit::Authorization
   before_action :authenticate_user!
 
   def index
@@ -26,50 +25,48 @@ class ParentsController < ApplicationController
   end
 
   def edit
-    @parent = find_and_authorize_parent
-  rescue ActiveRecord::RecordNotFound
-    redirect_to parents_path, alert: "Parent not found."
-  rescue Pundit::NotAuthorizedError
-    redirect_to parents_path, alert: "You are not authorized to edit this parent."
+    @parent = Parent.find(params[:id])
   end
 
   def update
-    @parent = find_and_authorize_parent
+    @parent = Parent.find(params[:id])
     if @parent.update(parent_params)
       redirect_to @parent, notice: "Parent information updated successfully."
     else
       render :edit
     end
-  rescue ActiveRecord::RecordNotFound
-    redirect_to parents_path, alert: "Parent not found."
-  rescue Pundit::NotAuthorizedError
-    redirect_to parents_path, alert: "You are not authorized to update this parent."
   end
 
-  def destroy
-    @parent = find_and_authorize_parent
+def destroy
+  @parent = Parent.find(params[:id])
+  authorize @parent
+
+  respond_to do |format|
     if @parent.destroy
-      redirect_to parents_path, notice: "Parent was successfully deleted."
+      format.html { redirect_back(fallback_location: root_path, notice: "Parent was successfully deleted.") }
+      format.turbo_stream { flash.now[:notice] = "Parent was successfully deleted." }
     else
-      redirect_to parents_path, alert: "Parent could not be deleted."
+      format.html { redirect_back(fallback_location: root_path, alert: "Parent could not be deleted.") }
+      format.turbo_stream { flash.now[:alert] = "Parent could not be deleted." }
     end
-  rescue ActiveRecord::RecordNotFound
-    redirect_to parents_path, alert: "Parent not found."
-  rescue Pundit::NotAuthorizedError
-    redirect_to parents_path, alert: "You are not authorized to delete this parent."
   end
+rescue ActiveRecord::RecordNotFound
+  redirect_to parents_path, alert: "Parent not found."
+rescue Pundit::NotAuthorizedError
+  redirect_to parents_path, alert: "You are not authorized to delete this parent."
+end
 
   def add_child
-    @parent = find_and_authorize_parent
-    @child = @parent.children.build
-  rescue ActiveRecord::RecordNotFound
-    redirect_to parents_path, alert: "Parent not found."
-  rescue Pundit::NotAuthorizedError
-    redirect_to parents_path, alert: "You are not authorized to add a child to this parent."
+    @parent = current_user.parent
+    if @parent.nil?
+      redirect_to new_parent_path, notice: "Please create a parent profile first."
+    else
+      @child = @parent.children.build
+    end
   end
 
   def create_child
-    @parent = find_and_authorize_parent
+    @parent = current_user.parent
     @child = @parent.children.new(child_params)
 
     if @child.save
@@ -77,27 +74,15 @@ class ParentsController < ApplicationController
     else
       render :add_child
     end
-  rescue ActiveRecord::RecordNotFound
-    redirect_to parents_path, alert: "Parent not found."
-  rescue Pundit::NotAuthorizedError
-    redirect_to parents_path, alert: "You are not authorized to add a child to this parent."
   end
 
   private
 
-  def find_and_authorize_parent
-    parent = Parent.find(params[:id])
-    authorize parent
-    parent
+  def child_params
+    params.require(:child).permit(:first_name, :last_name, :age, :grade, :food_allergies, :special_medical_needs, :emergency_contact)
   end
 
   def parent_params
-    params.require(:parent).permit(:first_name, :last_name, :phone_number, :email,
-      children_attributes: [:first_name, :last_name, :age, :grade,:food_allergies, :special_medical_needs, :emergency_contact])
-  end
-
-  def child_params
-    params.require(:child).permit(:first_name, :last_name, :age, :grade,
-      :food_allergies, :special_medical_needs, :emergency_contact)
+    params.require(:parent).permit(:first_name, :last_name, :phone_number, :email)
   end
 end
